che fine ha fatto il vecchio bionic.ts che ho cancellato?
