+++ src/classes/incubation/bionic.ts
@@ -12,7 +12,14 @@ import path = require('path')
 
 import { IRemix, IDistro } from '../../interfaces'
 
+import Fisherman from './fisherman'
+
+
 const exec = require('../../lib/utils').exec
+interface IReplaces {
+   search: string
+   replace: string
+}
 
 /**
  *
@@ -28,12 +35,21 @@ export class Bionic {
 
    user_opt: string
 
-   rootTemplate = './../../../conf/calamares/'
+   rootTemplate = path.resolve(__dirname, './../../../conf/calamares/bionic')
 
-   dirCalamaresModules = '/usr/lib/x86_64-linux-gnu/calamares/modules/'
+   dirCalamaresModules = '/usr/lib/calamares/modules/' // E DIFFERENTE in BIONIC
 
-   dirModules = '/etc/calamares/modules/'
+   dirModules = '/etc/calamares/modules'
 
+   fisherman = {}
+
+
+   /**
+    * @param remix
+    * @param distro
+    * @param displaymanager
+    * @param verbose
+    */
    constructor(remix: IRemix, distro: IDistro, displaymanager: boolean, user_opt: string, verbose = false) {
       this.remix = remix
       this.distro = distro
@@ -44,57 +60,55 @@ export class Bionic {
          this.dirCalamaresModules = '/usr/lib/calamares/modules/'
       }
       this.rootTemplate = path.resolve(__dirname, this.rootTemplate)
+
    }
 
    /**
     * write setting
     */
    settings() {
-      
+      const file = '/etc/calamares/settings.conf'
+      shx.cp(`${this.rootTemplate}/settings.conf`, '/etc/calamares')
+      shx.sed('-i', '%branding%', this.remix.branding, '/etc/calamares/settings.conf')
    }
+
 
    /**
     *
     */
-   modules() {
-      this.modulePartition()
-      this.moduleMount()
-      this.moduleUnpackfs()
-      this.moduleMachineid()
-      this.moduleFstab()
-      this.moduleLocale()
-      this.moduleKeyboard()
-      this.moduleLocalecfg()
-      this.moduleLuksbootkeyfile()
-      this.moduleUsers()
-      this.moduleDisplaymanager()
-      this.moduleNetworkcfg()
-      this.moduleHwclock()
+   async modules() {
+      const fisherman = new Fisherman(this.dirModules, this.dirCalamaresModules, this.rootTemplate, this.verbose)
+
+      await fisherman.buildModule('partition')
+      await fisherman.buildModule('mount')
+      await this.moduleUnpackfs()
+      await fisherman.buildModule('machineid')
+      await fisherman.buildModule('fstab')
+      await fisherman.buildModule('locale')
+      await fisherman.buildModule('keyboard')
+      await fisherman.buildModule('localecfg')
+      await fisherman.buildModule('luksbootkeyfile')
+      await fisherman.buildModule('users')
+      await this.moduleDisplaymanager()
+      await fisherman.buildModule('networkcfg')
+      await fisherman.buildModule('hwclock')
       this.moduleBeforebootloadermkdirs()
       this.moduleBug()
-      this.moduleInitramfscfg()
-      this.moduleInitramfs()
-      this.moduleGrubcfg()
+      await fisherman.buildModule('initramfscfg')
+      await fisherman.buildModule('initramfs')
+      await fisherman.buildModule('rubcfg')
       this.moduleBeforebootloader()
-      this.moduleBootloader()
+      await fisherman.buildModule('bootloader')
       this.moduleAfterbootloader()
-      this.moduleAutomirror()
       this.moduleAdd386arch()
       this.modulePackages()
       this.moduleRemoveuser()
-      this.moduleUmount()
-      this.moduleFinished()
+      await fisherman.buildCalamaresModule('remove-link', true)
+      // await fisherman.shellprocess('logs') non trova calamares-helper
+      await fisherman.buildModule('umount')
+      await fisherman.buildModule('finished')
    }
 
-   /**
-    * write module
-    * @param name
-    * @param content
-    */
-   private module(name: string, content: string) {
-      const file = this.dirModules + name + '.conf'
-      write(file, content, this.verbose)
-   }
 
    /**
     * ====================================================================================
@@ -119,7 +133,6 @@ export class Bionic {
          timeout: '600'
       })
       write(dir + 'module.desc', desBeforeBootloaderMkdirs)
-
       const bashContent = 'cp /lib/live/mount/medium/live/vmlinuz /boot/vmlinuz-$(uname -r)\n'
       const bashFile = `/usr/sbin/${name}`
       write(bashFile, bashContent, this.verbose)
@@ -238,238 +251,47 @@ export class Bionic {
       await exec(`chmod +x ${bashFile}`)
    }
 
-   /**
-    *
-    */
-   private modulePartition() {
-      const partition = yaml.safeDump({
-         efiSystemPartition: '/boot/efi',
-         enableLuksAutomatedPartitioning: true,
-         userSwapChoices: 'none',
-         drawNestedPartitions: true,
-         defaultFileSystemType: 'ext4'
-      })
-      this.module('partition', partition)
-   }
-
-   /**
-    *
-    */
-   private moduleMount() {
-      const mount = yaml.safeDump({
-         extraMounts: [
-            {
-               device: 'proc',
-               fs: 'proc',
-               mountPoint: '/proc'
-            },
-            {
-               device: 'sys',
-               fs: 'sysfs',
-               mountPoint: '/sys'
-            },
-            {
-               device: '/dev',
-               mountPoint: '/dev',
-               options: 'bind'
-            },
-            {
-               device: '/dev/pts',
-               fs: 'devpts',
-               mountPoint: '/dev/pts'
-            },
-            {
-               device: 'tmpfs',
-               fs: 'tmpfs',
-               mountPoint: '/run'
-            },
-            {
-               device: '/run/udev',
-               mountPoint: '/run/udev',
-               options: 'bind'
-            }
-         ],
-         extraMountsEfi: [
-            {
-               device: 'efivarfs',
-               fs: 'tmpefivarfsfs',
-               mountPoint: '/sys/firmware/efi/efivars'
-            }
-         ]
-      })
-
-      this.module('mount', mount)
-   }
 
    /**
     *
     */
    private moduleUnpackfs() {
-      const unpack = yaml.safeDump({
-         unpack: [
-            {
-               source: this.distro.mountpointSquashFs,
-               sourcefs: 'squashfs',
-               destination: ''
-            }
-         ]
-      })
-      this.module('unpackfs', unpack)
+      const fisherman = new Fisherman(this.dirModules, this.dirCalamaresModules, this.rootTemplate, this.verbose)
+      const name = 'unpackfs'
+      fisherman.buildModule(name)
+      shx.sed('-i', '%source%', this.distro.mountpointSquashFs, `${this.dirModules}/${name}.conf`)
    }
-
+   
    /**
-    *
+    * Al momento rimane con la vecchia configurazione
     */
-   private moduleMachineid() {
-      const machineid = yaml.safeDump({
-         systemd: true,
-         dbus: true,
-         symlink: true
-      })
-      this.module('machineid', machineid)
-   }
-
-   /**
-    *
-    */
-   private moduleFstab() {
-      const fstab = yaml.safeDump({
-         mountOptions: {
-            default: 'defaults,noatime',
-            btrfs: 'defaults,noatime,space_cache,autodefrag'
-         },
-         ssdExtraMountOptions: {
-            ext4: 'discard',
-            jfs: 'discard',
-            xfs: 'discard',
-            swap: 'discard',
-            btrfs: 'discard,compress=lzo'
-         },
-         crypttabOptions: 'luks,keyscript=/bin/cat'
-      })
-
-      this.module('fstab', fstab)
-   }
-   private moduleLocale() {
-      if (this.verbose) console.log(`calamares: module locale. Nothing to do!`)
-   }
-
-   private moduleKeyboard() {
-      if (this.verbose) console.log(`calamares: module keyboard. Nothing to do!`)
-   }
-
-   private moduleLocalecfg() {
-      if (this.verbose) console.log(`calamares: module localecfg. Nothing to do!`)
-   }
-
-   /**
-    *
-    */
-   private moduleUsers() {
-      const users = yaml.safeDump({
-         userGroup: 'users',
-         defaultGroups: ['cdrom', 'floppy', 'sudo', 'audio', 'dip', 'video', 'plugdev', 'netdev', 'lpadmin', 'scanner', 'bluetooth'],
-         autologinGroup: 'autologin',
-         sudoersGroup: 'sudo',
-         setRootPassword: false
-      })
-      this.module('users', users)
-   }
-
-   /**
-    *
-    */
-   private moduleDisplaymanager() {
-      const displaymanager_not_used = yaml.safeDump({
-         displaymanager: 'lightdm',
-         basicSetup: false,
-         sysconfigSetup: false
-      })
-
+   private async moduleDisplaymanager() {
+      const name = 'displaymanager'
       const displaymanager = require('./modules/displaymanager').displaymanager
-      this.module('displaymanager', displaymanager())
-   }
-
-   private moduleNetworkcfg() {
-      if (this.verbose) console.log(`calamares: module networkcfg. Nothing to do!`)
-   }
-
-   private moduleHwclock() {
-      if (this.verbose) console.log(`calamares: module hwclock. Nothing to do!`)
-   }
-
-   private moduleServicesSystemd() {
-      if (this.verbose) console.log(`calamares: module servives-systemd. Nothing to do!`)
-   }
-
-   private moduleGrubcfg() {
-      if (this.verbose) console.log(`calamares: module grubcfg. Nothing to do!`)
+      const file = this.dirModules + name + '.conf'
+      const content = displaymanager()
+      fs.writeFileSync(file, content, 'utf8')
    }
 
    /**
-    *
+    * Al momento rimane con la vecchia configurazione
     */
-   private moduleBootloader() {
-      const bootloader = yaml.safeDump({
-         efiBootLoader: 'grub',
-         kernel: '/vmlinuz-linux',
-         img: '/initramfs-linux.img',
-         fallback: '/initramfs-linux-fallback.img',
-         timeout: 10,
-         grubInstall: 'grub-install',
-         grubMkconfig: 'grub-mkconfig',
-         grubCfg: '/boot/grub/grub.cfg',
-         grubProbe: 'grub-probe',
-         efiBootMgr: 'efibootmgr',
-         installEFIFallback: false
-      })
-      this.module('bootloader', bootloader)
-   }
-
-   /**
-    * create module packages.conf
-    */
-   private modulePackages() {
+   private async modulePackages() {
       const packages = require('./modules/packages').packages
-      this.module('packages', packages())
+      const content = packages()
+      const name = 'packages'
+      const file = this.dirModules + name + '.conf'
+      fs.writeFileSync(file, content, 'utf8')
    }
 
-   private moduleLuksbootkeyfile() {
-      if (this.verbose) console.log(`calamares: module luksbootkeyfile. Nothing to do!`)
-   }
-
-   /**
-    *
+      /**
+    * Al momento rimane con la vecchia configurazione
     */
-   private module_luksopenswaphookcfg() {
-      const lksopenswaphookcfg = yaml.safeDump({
-         configFilePath: '/etc/openswap.conf'
-      })
-      this.module('lksopenswaphookcfg', lksopenswaphookcfg)
-   }
-
-   private modulePlymouthcfg() {
-      if (this.verbose) console.log(`calamares: module plymouthcfg. Nothing to do!`)
-   }
-
-   private moduleInitramfscfg() {
-      if (this.verbose) console.log(`calamares: module initramfscfg. Nothing to do!`)
-   }
-
-   /**
-    *
-    */
-   private moduleRemoveuser() {
-      const removeuser = yaml.safeDump({ username: this.user_opt })
-      this.module('removeuser', removeuser)
-   }
-
-   private moduleInitramfs() {
-      if (this.verbose) console.log(`calamares: module initramfs. Nothing to do!`)
-   }
-
-   private moduleUmount() {
-      if (this.verbose) console.log(`calamares: module unmount. Nothing to do!`)
+   private async moduleRemoveuser() {
+      const name = 'removeuser'
+      const content = yaml.safeDump({ username: this.user_opt })
+      const file = this.dirModules + name + '.conf'
+      fs.writeFileSync(file, content, 'utf8')
    }
 
    /**
@@ -491,69 +313,7 @@ export class Bionic {
     * ====================================================================================
     */
 
-   /**
-    * Automirror
-    * Pythonm
-    */
-   private async moduleAutomirror() {
-      const name = 'automirror'
-      const dirModule = this.dirCalamaresModules + name + '/'
-      if (!fs.existsSync(dirModule)) {
-         fs.mkdirSync(dirModule)
-      }
 
-      const automirror = yaml.safeDump({
-         baseUrl: 'archive.ubuntu.com',
-         distribution: 'Ubuntu',
-         geoip: {
-            style: 'json',
-            url: 'http  s://ipapi.co/json'
-         }
-      })
-      write(dirModule + 'automirror.conf', automirror, this.verbose)
-
-      // Creo anche un config in local con la distro particolare, esempio: lubuntu, ulyana
-      const automirrorModules = yaml.safeDump({
-         baseUrl: 'archive.ubuntu.com',
-         distribution: 'Lubuntu',
-         geoip: {
-            style: 'json',
-            url: 'https://ipapi.co/json'
-         }
-      })
-      write('/etc/calamares/modules/' + 'automirror.conf', automirrorModules)
-
-      // desc
-      const desc = yaml.safeDump({
-         type: 'job',
-         name: 'automirror',
-         interface: 'python',
-         script: 'main.py'
-      })
-      write(dirModule + 'module.desc', desc, this.verbose)
-
-      // py
-      const scriptAutomirror = require('./calamares-modules/scripts/automirror').automirror
-      const scriptFile = dirModule + 'main.py'
-      write(scriptFile, scriptAutomirror(this.distro.versionId), this.verbose)
-      await exec(`chmod +x ${scriptFile}`)
-   }
-
-   private async moduleCreatetmp() {
-      const name = 'create-tmp'
-      const dirModule = this.dirCalamaresModules + name + '/'
-      if (!fs.existsSync(dirModule)) {
-         fs.mkdirSync(dirModule)
-      }
-
-      const createTmp = require('./calamares-modules/desc/create-tmp').createTmp
-      write(dirModule + 'module.desc', createTmp(), this.verbose)
-
-      const scriptcreateTmp = require('./calamares-modules/scripts/create-tmp').createTmp
-      const scriptFile = `/usr/sbin/${name}`
-      write(scriptFile, scriptcreateTmp(), this.verbose)
-      await exec(`chmod +x ${scriptFile}`)
-   }
 
    /**
     *
@@ -573,6 +333,12 @@ export class Bionic {
       write(scriptFile, scriptBootloaderConfig(), this.verbose)
       await exec(`chmod +x ${scriptFile}`)
    }
+
+   private module(name: string, content: string) {
+      const file = this.dirModules + name + '.conf'
+      write(file, content, this.verbose)
+   }
+
 }
 
 /**
